added CheckAccessHelper component, added AppUserResponseGetPayload.

+++ src/main/java/ru/cft/shiftbanquet/controller/ExpenseController.java
@@ -1,9 +1,8 @@
 package ru.cft.shiftbanquet.controller;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
-import io.swagger.annotations.ApiParam;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import ru.cft.shiftbanquet.entity.Event;
 import ru.cft.shiftbanquet.entity.Expense;
@@ -11,7 +10,7 @@ import ru.cft.shiftbanquet.entity.Wrapper;
 import ru.cft.shiftbanquet.payloads.ExpenseRequestPostPayload;
 import ru.cft.shiftbanquet.repos.EventRepo;
 import ru.cft.shiftbanquet.repos.ExpenseRepo;
-import sun.reflect.generics.reflectiveObjects.NotImplementedException;
+import ru.cft.shiftbanquet.service.CheckAccessHelper;
 
 import java.util.List;
 
@@ -25,10 +24,18 @@ public class ExpenseController {
     @Autowired
     private EventRepo eventRepo;
 
+    @Autowired
+    private CheckAccessHelper accessHelper;
+
     @ApiOperation(value = "Редактировать расходы")
     @PutMapping("api/events/{event_id}/expenses/{expense_id}")
     public Wrapper<Expense> editExpense(@PathVariable long event_id, @PathVariable long expense_id, @RequestBody Wrapper<ExpenseRequestPostPayload> payload) {
         Event event = eventRepo.findEventById(event_id);
+
+        if (accessHelper.checkAccess(event)){
+            return new Wrapper<>("NO ACCESS", null);
+        }
+
         Expense expense = expenseRepo.findExpenseByIdAndEvent(expense_id, event);
         expense.setExpense(payload.getData());
         expenseRepo.save(expense);
@@ -38,8 +45,13 @@ public class ExpenseController {
     @ApiOperation(value = "Удалить расходы")
     @DeleteMapping("api/events/{event_id}/expenses/{expense_id}")
     public Wrapper<Expense> deleteExpense(@PathVariable long event_id, @PathVariable long expense_id) {
+
         Event event = eventRepo.findEventById(event_id);
         Expense expense = expenseRepo.findExpenseByIdAndEvent(expense_id, event);
+
+        if (accessHelper.checkAccess(event)){
+            return new Wrapper<>("NO ACCESS", null);
+        }
 
         if(expense != null) {
             expenseRepo.delete(expense);
@@ -55,6 +67,11 @@ public class ExpenseController {
     public Wrapper<Expense> addExpense(@PathVariable long event_id, @RequestBody Wrapper<ExpenseRequestPostPayload> payload) {
         ExpenseRequestPostPayload data = payload.getData();
         Event event = eventRepo.findEventById(event_id);
+
+        if (accessHelper.checkAccess(event)){
+            return new Wrapper<>("NO ACCESS", null);
+        }
+
         Expense expense = new Expense();
         expense.setEvent(event);
         expense.setCost(data.getCost());
@@ -75,4 +92,5 @@ public class ExpenseController {
         Event event = eventRepo.findEventById(event_id);
         return new Wrapper<>("OK", expenseRepo.findExpenseByIdAndEvent(expense_id, event));
     }
+
 }
